+++ packages/app/src/components-styled/data-driven-text.tsx
@@ -4,22 +4,25 @@ import { National } from "~/types/data";
 import { assert } from "~/utils/assert";
 import { formatNumber } from "~/utils/formatNumber";
 import { getLastFilledValue } from "~/utils/get-last-filled-value";
-import { getPluralizedText, PluralizationTexts } from "~/utils/get-pluralized-text";
 import { replaceComponentsInText } from "~/utils/replace-components-in-text";
 import { DifferenceIndicator } from "./difference-indicator";
 import { RelativeDate } from "./relative-date";
+import { Text } from "./typography";
 
 interface DataDrivenTextProps {
   data: National;
   differenceKey: string;
-  metricName: string;
+  metricName: keyof National;
   metricProperty: string;
-  baseTexts: PluralizationTexts;
-  additionalTexts: PluralizationTexts;
+  valueTexts: PluralizationTexts;
+  differenceTexts: PluralizationTexts;
 }
 
-export function DataDrivenText({ data, differenceKey, metricName, metricProperty, baseTexts, additionalTexts }: DataDrivenTextProps) {
+export function DataDrivenText({
+  data, differenceKey, metricName, metricProperty, valueTexts, differenceTexts
+}: DataDrivenTextProps) {
 
+  // @Todo move to a lastValue abstraction
   const lastValue =
     metricProperty === 'hospital_moving_avg_per_region'
       ? getLastFilledValue(data, metricName)
@@ -39,7 +42,7 @@ export function DataDrivenText({ data, differenceKey, metricName, metricProperty
   );
 
   const differenceValue = differenceKey
-    ? get(data, ['difference', (differenceKey as unknown) as string])
+    ? get(data, ['difference', differenceKey])
     : undefined;
 
   assert(
@@ -47,10 +50,10 @@ export function DataDrivenText({ data, differenceKey, metricName, metricProperty
     `Missing value for difference:${differenceKey}`
   );
 
-  const baseText = getPluralizedText(baseTexts, propertyValue);
-  const additionalText = getPluralizedText(additionalTexts, differenceValue.difference);
+  const baseText = getPluralizedText(valueTexts, propertyValue);
+  const additionalText = getPluralizedText(differenceTexts, differenceValue.difference);
 
-  return <p>
+  return <Text>
     {replaceComponentsInText(baseText, {
       newDate: <RelativeDate dateInSeconds={differenceValue.new_date_of_report_unix} isCapitalized />,
       propertyValue: <strong>{formatNumber(propertyValue)}</strong>
@@ -60,5 +63,19 @@ export function DataDrivenText({ data, differenceKey, metricName, metricProperty
       differenceIndicator: <DifferenceIndicator value={differenceValue} context="inline" />,
       oldDate: <RelativeDate dateInSeconds={differenceValue.old_date_of_report_unix} />
     })}
-  </p>;
+  </Text>;
 }
+
+type PluralizationTexts = Record<'zero' | 'singular' | 'plural', string>;
+
+function getPluralizedText(texts: PluralizationTexts, count: number): string {
+  const absoluteCount = Math.abs(count);
+  if (absoluteCount === 0) {
+    return texts.zero;
+  }
+  if (absoluteCount <= 1) {
+    return texts.singular;
+  }
+  return texts.plural;
+}
+
