+++ packages/app/src/locale/en.json
@@ -759,12 +759,12 @@
     "data_sitemap_toelichting": "",
     "data_driven_texts": {
       "infected_people_total": {
-        "base": {
+        "value": {
           "zero": "{{newDate}} there were no newly reported infected people.",
           "singular": "{{newDate}} there was {{propertyValue}} newly reported infected person.",
           "plural": "{{newDate}} there were {{propertyValue}} newly reported infected people."
         },
-        "additional": {
+        "difference": {
           "zero": "This is the same as {{oldDate}}",
           "singular": "This is {{differenceIndicator}} than {{oldDate}}.",
           "plural": "This is {{differenceIndicator}} than {{oldDate}}."
