+++ packages/app/src/pages/index.tsx
@@ -41,8 +41,8 @@ const Home: FCWithLayout<IHomeData> = (data) => {
         metricName="infected_people_total"
         metricProperty="infected_daily_total"
         differenceKey="infected_people_total__infected_daily_total"
-        baseTexts={siteText.nationaal_actueel.data_driven_texts.infected_people_total.base}
-        additionalTexts={siteText.nationaal_actueel.data_driven_texts.infected_people_total.additional} />
+        valueTexts={text.data_driven_texts.infected_people_total.value}
+        differenceTexts={text.data_driven_texts.infected_people_total.difference} />
       <ChoroplethTile
         title={text.risiconiveaus.selecteer_titel}
         description={
